move history logic to its own reducer
