+++ src/components/App/Dashboard/Balances/BalancesContainer.js
@@ -19,7 +19,6 @@ const mapStateToProps = state => ({
       buyAsset: 'usd',
     }),
   ],
-  history: state.prices.history,
 });
 
 const actions = {
