+++ src/components/shared/Chart/index.js
@@ -1,1 +1,1 @@
-export { default } from './Chart';
+export { default } from './ChartContainer';
