+++ src/state/history/actions.js
@@ -1,0 +1,3 @@
+import { createActions } from 'redux-yo';
+
+export const historyActions = createActions(['setHistory']);
